Split TaskManager ControllerProvider into 2 classes
